backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.6.1
httpx==0.25.2
tenacity==8.2.3
redis==5.0.1
PyYAML==6.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
playwright==1.40.0
pytest==7.4.3
allure-pytest==2.13.2